refactor(BorrowerSideSheet): move side-sheet out of BorrowerProfile
BorrowerSideSheetWrapper/Content and the SideSheet* sub-components are a
KvSideSheet drawer triggered from loan cards, not borrower-profile page
components. Move their source and Storybook stories into a dedicated
BorrowerSideSheet directory and repoint imports.

AD-281
