+++ src/components/BorrowerSideSheet/BorrowerSideSheetContent.vue
@@ -84,13 +84,13 @@ import { fetchAiLoanPills } from '#src/util/aiLoanPIillsUtils';
 import { addMonths, differenceInWeeks } from 'date-fns';
 import { FUNDRAISING, ENDED } from '#src/api/fixtures/LoanStatusEnum';
 import LoanNextSteps from '#src/components/Thanks/LoanNextSteps';
-import CommentsAndWhySpecial from './CommentsAndWhySpecial';
-import BorrowerCountry from './BorrowerCountry';
-import ContributingPartners from './ContributingPartners';
-import DetailsTabs from './DetailsTabs';
-import LendersAndTeams from './LendersAndTeams';
-import LoanProgress from './LoanProgress';
-import MoreAboutLoan from './MoreAboutLoan';
+import CommentsAndWhySpecial from '#src/components/BorrowerProfile/CommentsAndWhySpecial';
+import BorrowerCountry from '#src/components/BorrowerProfile/BorrowerCountry';
+import ContributingPartners from '#src/components/BorrowerProfile/ContributingPartners';
+import DetailsTabs from '#src/components/BorrowerProfile/DetailsTabs';
+import LendersAndTeams from '#src/components/BorrowerProfile/LendersAndTeams';
+import LoanProgress from '#src/components/BorrowerProfile/LoanProgress';
+import MoreAboutLoan from '#src/components/BorrowerProfile/MoreAboutLoan';
 import SideSheetHeader from './SideSheetHeader';
 import SideSheetLoanHowMoneyHelps from './SideSheetLoanHowMoneyHelps';
 import SideSheetLoanStory from './SideSheetLoanStory';
