+++ src/components/BorrowerSideSheet/BorrowerSideSheetWrapper.vue
@@ -22,7 +22,7 @@
 </template>
 
 <script setup>
-import BorrowerSideSheetContent from '#src/components/BorrowerProfile/BorrowerSideSheetContent';
+import BorrowerSideSheetContent from '#src/components/BorrowerSideSheet/BorrowerSideSheetContent';
 import { KvSideSheet } from '@kiva/kv-components';
 
 const props = defineProps({
